Vital subsystem stable and returning expected values.

+++ PVMSBedsideMonitor/src/vital/BodyTemperature.java
@@ -10,13 +10,13 @@ public class BodyTemperature implements VitalAlgorithm {
 	
 	
 	@Override
-	public int getValue(int rawMeasure) {
-		int realValue = MIN + ((MAX - MIN) * (rawMeasure/100) );
+	public int getValue(double rawMeasure) {
+		int realValue = (int) (MIN + ((MAX - MIN) * (rawMeasure/100) ));
 		return realValue;
 	}
-	
+
 	public int getValue(){
-		int realValue = sensor.read();
+		double realValue = sensor.read();
 		return getValue(realValue);
 	}
 }
