+++ PVMSBedsideMonitor/src/vital/RespiratoryRate.java
@@ -9,13 +9,13 @@ public class RespiratoryRate implements VitalAlgorithm {
 	VitalSensor sensor = new VitalSensor("respiratoryrate");
 	
 	@Override
-	public int getValue(int rawMeasure) {
-		int realValue = MIN + ((MAX - MIN) * (rawMeasure/100) );
+	public int getValue(double rawMeasure) {
+		int realValue = (int) (MIN + ((MAX - MIN) * (rawMeasure/100) ));
 		return realValue;
 	}
 
 	public int getValue(){
-		int realValue = sensor.read();
+		double realValue = sensor.read();
 		return getValue(realValue);
 	}
